EEPROM library: move function definitions into separate .cpp files
